Int32の判定とfloatの判定が混在
git-svn-id: file:///home/moriyoshi/chaperone.e-cell.org/var/opt/subversion/ecell-ide/trunk@3255 fe1fbaf7-ad0a-4b3d-a902-8cb9c273ac6b

+++ ide/CBGridLayout/CBGridLayoutPanel.cs
@@ -273,6 +273,11 @@ namespace Ecell.IDE.Plugins.CBGridLayout
             }
         }
 
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="sender"></param>
+        /// <param name="e"></param>
         private void marginTextBox_Validating(object sender, System.ComponentModel.CancelEventArgs e)
         {
             CBGridLayout alogorithm = (CBGridLayout)m_algorithm;
@@ -284,8 +289,8 @@ namespace Ecell.IDE.Plugins.CBGridLayout
                 e.Cancel = true;
                 return;
             }
-            float dummy;
-            if (!float.TryParse(text, out dummy) || dummy < 0.0)
+            int dummy;
+            if (!Int32.TryParse(text, out dummy) || dummy < 0)
             {
                 Util.ShowErrorDialog(string.Format(MessageResGridLayout.ErrInvalidValue, label5.Text));
                 marginTextBox.Text = Convert.ToString(alogorithm.Margin);
