+++ ide/GridLayout/GridLayout/GridLayoutPanel.cs
@@ -198,8 +198,8 @@ namespace Ecell.IDE.Plugins.GridLayout
                 e.Cancel = true;
                 return;
             }
-            float dummy;
-            if (!float.TryParse(text, out dummy) || dummy < 0.0)
+            int dummy;
+            if (!Int32.TryParse(text, out dummy) || dummy < 0)
             {
                 Util.ShowErrorDialog(string.Format(MessageResGridLayout.ErrInvalidValue, defMargin.Text));
                 defMargin.Text = Convert.ToString(alogorithm.DefMargin);
@@ -223,8 +223,8 @@ namespace Ecell.IDE.Plugins.GridLayout
                 e.Cancel = true;
                 return;
             }
-            float dummy;
-            if (!float.TryParse(text, out dummy) || dummy < 0.0)
+            int dummy;
+            if (!Int32.TryParse(text, out dummy) || dummy < 0)
             {
                 Util.ShowErrorDialog(string.Format(MessageResGridLayout.ErrInvalidValue, annealingRepeats.Text));
                 annealingRepeats.Text = Convert.ToString(alogorithm.Kmax);
@@ -249,8 +249,8 @@ namespace Ecell.IDE.Plugins.GridLayout
                 e.Cancel = true;
                 return;
             }
-            float dummy;
-            if (!float.TryParse(text, out dummy) || dummy < 0.0)
+            int dummy;
+            if (!Int32.TryParse(text, out dummy) || dummy < 0)
             {
                 Util.ShowErrorDialog(string.Format(MessageResGridLayout.ErrInvalidValue, initialT.Text));
                 initialT.Text = Convert.ToString(alogorithm.InitialT);
@@ -275,8 +275,8 @@ namespace Ecell.IDE.Plugins.GridLayout
                 e.Cancel = true;
                 return;
             }
-            float dummy;
-            if (!float.TryParse(text, out dummy) || dummy < 0.0)
+            int dummy;
+            if (!Int32.TryParse(text, out dummy) || dummy < 0)
             {
                 Util.ShowErrorDialog(string.Format(MessageResGridLayout.ErrInvalidValue, naturalLength.Text));
                 naturalLength.Text = Convert.ToString(alogorithm.NaturalLength);
